% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/binomial.R
\name{binomial}
\alias{binomial}
\title{Binomial test for allele specific expression in a single sample}
\usage{
binomial(data_mat)
}
\arguments{
\item{data}{matrix with first column one of the allele counts and the second column total allele \code{inputParameter1}}

\item{inputParameter2}{A description of the input parameter \code{inputParameter2}}
}
\value{
output A description of the object the function outputs
}
\description{
A more detailed description of what the function is and how
it works. It may be a paragraph that should not be separated
by any spaces.
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

