% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/beta_binomial.R, R/dbetabinom.R
\name{dbetabinom}
\alias{dbetabinom}
\title{Beta-Binomial test for testing allele specific expression
using multiple samples}
\usage{
dbetabinom(alpha, beta, hi, tot)

dbetabinom(alpha, beta, hi, tot)
}
\arguments{
\item{inputParameter1}{A description of the input parameter \code{inputParameter1}}

\item{inputParameter2}{A description of the input parameter \code{inputParameter2}}

\item{data}{matrix with first column one of the allele counts and the second column total allele \code{inputParameter1}}

\item{inputParameter2}{A description of the input parameter \code{inputParameter2}}
}
\value{
output A description of the object the function outputs

output A description of the object the function outputs
}
\description{
A more detailed description of what the function is and how
it works. It may be a paragraph that should not be separated
by any spaces.

A more detailed description of what the function is and how
it works. It may be a paragraph that should not be separated
by any spaces.
}
\examples{
R code here showing how your function works
R code here showing how your function works
}
\keyword{keywords}

