% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/log_liklihood_h0.R
\name{log_liklihood_h0}
\alias{log_liklihood_h0}
\title{R function to compute log liklihood for Beta Binomial model under null hypothesis}
\usage{
log_liklihood_h0(alpha, his, tots)
}
\arguments{
\item{data}{matrix with first column one of the allele counts and the second column total allele \code{inputParameter1}}

\item{inputParameter2}{A description of the input parameter \code{inputParameter2}}
}
\value{
output A description of the object the function outputs
}
\description{
A more detailed description of what the function is and how
it works. It may be a paragraph that should not be separated
by any spaces.
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

